+++ Pikomon API/src/main/java/br/com/pikomon/Pikomon/dto/CreatePokemonDTO.java
@@ -1,4 +1,4 @@
 package br.com.pikomon.Pikomon.dto;
 
-public record CreatePokemonDTO(int id,int level,String trainer) {
+public record CreatePokemonDTO(Long id,int level,String trainer) {
 }
